gitgoodrentals/backend/requirements.txt
asgiref==3.7.2
astroid==3.1.0
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-google-maps==0.13.0
djangorestframework==3.15.0
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.10.0
